WIP: Extend DFU to support bulk transfers
Extend the protocol to add a new DFU_BULK_DOWNLOAD message
that instructs the device to expect bulk OUT messages that
are N times 4K blocks in size.
The number of blocks is specified in wValue. The remaining
unaligned data is sent using the normal DNLOAD commands.

This change aims to allow DFU to be used more efficiently
on devices with large ammounts of flash storage (EMMC) or to
or for loading Linux + initrd into memory.

TODO
Update descriptors / versions to indicate bulk x-fer support
plus a maximum 'bulk' xfer chunk size e.g. the size of the
EMMC cached write buffering.